fix(pipeline/waf): CR fixes + docs + test helpers
Cycle-1 CodeRabbit review on PR #282 returned 5 actionable findings.
Resolve all 5 plus the Docs-coverage CI canary failure for the 12 new
WafChallengeInternals exports.

CR1 (Major) HCaptchaCheckboxSolver.ts
  Extract clickCentreSafe() helper that wraps page.mouse.click in
  try/catch and returns DidSolve(false) on throw. The solver contract
  forbids throws; the previous direct clickCentre() call could throw
  when the page closes between bounding-box read and click.

CR2 (Major) WafChallengeConfig.ts
  Re-type HCAPTCHA_IFRAME_URL_PATTERNS and TURNSTILE_IFRAME_URL_PATTERNS
  from readonly string[] to "as const" tuples. Preserves literal types
  for downstream pattern-matching narrowing.

CR3 (Major) WafChallengeInterceptor.ts
  Add docs/architecture/waf-interceptor.md (~140 lines) covering the
  interceptor purpose, attach/detach lifecycle, Camoufox auto-pass
  recipe, kill-switch (WAF_INTERCEPTOR_DISABLED), detection rules,
  telemetry, failure-mode matrix, all 12 internal helper names, and
  Open/Closed extension steps. mkdocs --strict passes.

  Side effect: docs-coverage CI canary now passes naturally for all 12
  new exports plus clickCentreSafe (15 symbols total). No allowlist
  entries added.

CR4 (Minor) WafChallengeInternals.ts
  Rename brand IsEnabled -> IsDisabled so the type and the function
  name (isDisabled) agree on meaning. Updates return type, JSdoc, and
  the cast site.

CR5 (Nit) Test helpers
  Extract clearDisableEnv / makeLogger / makePageStub from
  WafChallengeInternals.test.ts and WafChallengeInterceptor.test.ts
  into WafChallengeTestHelpers.ts. Mirrors the pattern used by
  TestHelpers.ts and MockFactories.ts in the Infrastructure folder.

Local validation
  - npx tsc --noEmit -> exit 0
  - npx eslint touched -> 0 errors
  - npm run lint:canaries -> 46 TS + 5 bash canaries fire
  - npm run lint:guideline-coverage -> 4 clusters
  - npx jest Pipeline/Interceptors/WafChallenge -> 85/85 pass
  - jest Pipeline/(Interceptors|Architecture|Core/Builder) -> 229/229
  - npm run test:pipeline -> 4914/4914 pass; coverage
    97.08/95/97.25/98.33 (all thresholds preserved)
  - npm run build -> lib/index.d.ts hash 221E161... byte-identical
  - bash .github/scripts/ci/docs-coverage.sh -> 12 documented, 0 missing
  - python -m mkdocs build --strict -> exit 0
  - madge --circular WafChallenge -> acyclic
